fix: Make consecutive 3xx (and 4xx) effects work
Only reset the full effect state on new note trigger.

This makes effect sequences like the following work:

301
302
463

In particular, when crossing a pattern boundary, the
following effect sequence might be observed:

301 // pattern break after this row
301 // same effect is repeated on the next pattern

Resetting the entire effect state caused the portamento
control flags and the target period value to be lost.
Keep in mind that the effect state is a union (memory
shared between all kinds of effect).

When setting a vibrato effect, the delay must be set to
zero to ensure that the vibrato is applied right away.